requirements.txt
Flask
python-nmap
pandas
plotly